requirements.txt
click>=8.1
tabulate>=0.9
streamlit>=1.35
google-auth>=2.20
google-api-python-client>=2.100
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.2
